IMR-155 refactor : split request, reponse, model dtos
