+++ back/src/dto/model/__init__.py
@@ -1,0 +1,7 @@
+from .artist import Artist
+from .inference import Inference, SongsElement
+from .playlist import Playlist
+from .song import Song
+from .user import User
+
+__all__ = [Artist, Inference, SongsElement, Playlist, Song, User]
