+++ back/src/dto/model/artist.py
