+++ back/src/dto/request/__init__.py
@@ -1,0 +1,4 @@
+from .feedback import UserFeedbackRequest
+from .music import RecommendMusicRequest
+
+__all__ = [UserFeedbackRequest, RecommendMusicRequest]
