+++ back/src/dto/response/__init__.py
@@ -1,0 +1,3 @@
+from .music import RecommendMusicResponse, RecommendMusic
+
+__all__ = [RecommendMusic, RecommendMusicResponse]
