+++ back/src/dto/response/music.py
@@ -1,5 +1,4 @@
 from pydantic import BaseModel
-from fastapi import Form
 
 
 class RecommendMusic(BaseModel):
@@ -13,12 +12,3 @@ class RecommendMusic(BaseModel):
 class RecommendMusicResponse(BaseModel):
     session_id: str
     songs: list[RecommendMusic]
-
-
-class RecommendMusicRequest(BaseModel):
-    genres: list[str]
-
-    @classmethod
-    def as_form(cls, genres: list[str] = Form()):
-        genres_ = genres[0].split(",")
-        return cls(genres=genres_)
